Dump Added

+++ Task3/SafeStackCStyle/Tests/SafeStackTester/SafeStackTest.cpp
@@ -9,23 +9,23 @@
 
 
 TEST(SafeStackTest, FULL_NESS) {
-    auto Stack = SafeStack<int>(-1);
+    auto Stack = SafeStack<size_t >(-1);
     for (size_t i = 0; i < 10000000; i++) {
         auto error_code = Stack.push(i);
-        if (error_code != SafeStack<int>::OK) {
-            ASSERT_EQ(SafeStack<int>::ErrorPush, error_code);
+        if (error_code != SafeStack<size_t >::OK) {
+            ASSERT_EQ(SafeStack<size_t >::ErrorPush, error_code);
             break;
         }
     }
 }
 
 TEST(SafeStackTest, CHANGE_FIRST_BYTES) {
-    auto Stack = SafeStack<int>(-1);
+    auto Stack = SafeStack<size_t >(-1);
     ((int *)(&Stack))[0] = 0;
     for (size_t i = 0; i < 10000000; i++) {
         auto error_code = Stack.push(i);
-        if (error_code != SafeStack<int>::OK) {
-            ASSERT_EQ(SafeStack<int>::ErroKanareika, error_code);
+        if (error_code != SafeStack<size_t >::OK) {
+            ASSERT_EQ(SafeStack<size_t >::ErroKanareika, error_code);
             break;
         }
     }
